refactor: removed AlternativeMods folder
