--- AlternativeMods/translate_MOD.js
@@ -1,178 +1,0 @@
-module.exports = {
-
-	//---------------------------------------------------------------------
-	// Action Name
-	//
-	// This is the name of the action displayed in the editor.
-	//---------------------------------------------------------------------
-
-	name: "Translate",
-
-	//---------------------------------------------------------------------
-	// Action Section
-	//
-	// This is the section the action will fall into.
-	//---------------------------------------------------------------------
-
-	section: "Other Stuff",
-
-	//---------------------------------------------------------------------
-	// Action Subtitle
-	//
-	// This function generates the subtitle displayed next to the name.
-	//---------------------------------------------------------------------
-
-	subtitle: function (data) {
-		return `Translate to [${data.translateTo}]`;
-	},
-
-	//---------------------------------------------------------------------
-	// DBM Mods Manager Variables (Optional but nice to have!)
-	//
-	// These are variables that DBM Mods Manager uses to show information
-	// about the mods for people to see in the list.
-	//---------------------------------------------------------------------
-
-	// Who made the mod (If not set, defaults to "DBM Mods")
-	author: "EGGSY",
-
-	// The version of the mod (Defaults to 1.0.0)
-	version: "1.8.6",
-
-	// A short description to show on the mod line for this mod (Must be on a single line)
-	short_description: "Translate words!",
-
-	// If it depends on any other mods by name, ex: WrexMODS if the mod uses something from WrexMods
-
-	//---------------------------------------------------------------------
-	// Action Storage Function
-	//
-	// Stores the relevant variable info for the editor.
-	//---------------------------------------------------------------------
-
-	variableStorage: function (data, varType) {
-		const type = parseInt(data.storage);
-		if (type !== varType) return;
-		let dataType = 'Translated String';
-		return ([data.varName, dataType]);
-	},
-
-	//---------------------------------------------------------------------
-	// Action Fields
-	//
-	// These are the fields for the action. These fields are customized
-	// by creating elements with corresponding IDs in the HTML. These
-	// are also the names of the fields stored in the action's JSON data.
-	//---------------------------------------------------------------------
-
-	fields: ["translateTo", "translateMessage", "storage", "varName"],
-
-	//---------------------------------------------------------------------
-	// Command HTML
-	//
-	// This function returns a string containing the HTML used for
-	// editting actions. 
-	//
-	// The "isEvent" parameter will be true if this action is being used
-	// for an event. Due to their nature, events lack certain information, 
-	// so edit the HTML to reflect this.
-	//
-	// The "data" parameter stores constants for select elements to use. 
-	// Each is an array: index 0 for commands, index 1 for events.
-	// The names are: sendTargets, members, roles, channels, 
-	//                messages, servers, variables
-	//---------------------------------------------------------------------
-
-	html: function (isEvent, data) {
-		return `
-<div>
-	<div id="modinfo" style="float: left;">
-	<p>
-	   <u>Mod Info:</u><br>
-	   Made by EGGSY!<br>
-	</p></div>
-	<div style="float: right; width: 60%;">
-		Translate to:<br>
-		<input id="translateTo" placeholder="Should be 2 letters." class="round" type="text" maxlength="2"><br>
-	</div>
-</div><br><br><br>
-<div style="padding-top: 8px;">
-	Translate Message:<br>
-	<textarea id="translateMessage" rows="9" placeholder="Insert message that you want to translate here..." style="width: 99%; font-family: monospace; white-space: nowrap; resize: none;"></textarea>
-</div><br>
-<div>
-	<div style="float: left; width: 35%;">
-		Store In:<br>
-		<select id="storage" class="round" onchange="glob.variableChange(this, 'varNameContainer')">
-			${data.variables[0]}
-		</select>
-	</div>
-	<div id="varNameContainer" style="display: none; float: right; width: 60%;">
-		Variable Name:<br>
-		<input id="varName" class="round" type="text">
-	</div>
-</div>`;
-	},
-
-	//---------------------------------------------------------------------
-	// Action Editor Init Code
-	//
-	// When the HTML is first applied to the action editor, this code
-	// is also run. This helps add modifications or setup reactionary
-	// functions for the DOM elements.
-	//---------------------------------------------------------------------
-
-	init: function () {
-		const {glob, document} = this;
-
-		glob.variableChange(document.getElementById('storage'), 'varNameContainer');
-	},
-
-	//---------------------------------------------------------------------
-	// Action Bot Function
-	//
-	// This is the function for the action within the Bot's Action class.
-	// Keep in mind event calls won't have access to the "msg" parameter, 
-	// so be sure to provide checks for variable existance.
-	//---------------------------------------------------------------------
-
-	action: function (cache) {
-
-		var _this = this;
-		const data = cache.actions[cache.index];
-		const translateTo = this.evalMessage(data.translateTo, cache);
-		const translateMessage = this.evalMessage(data.translateMessage, cache);
-		const storage = parseInt(data.storage);
-		const varName = this.evalMessage(data.varName, cache);
-
-		// Check if everything is ok
-		if (!translateTo || translateTo.length > 2) return console.log("Translate to can only be 2 letters.");
-		if (!translateMessage) return console.log("You need to write something to translate.");
-
-		// Main code
-		var WrexMODS = this.getWrexMods(); // Using Wrex mods, as always.
-		const translate = WrexMODS.require('node-google-translate-skidz');
-
-		translate({
-			text: translateMessage,
-			target: translateTo
-		}, function (result) {
-				if(result.translation !== undefined) {
-					_this.storeValue(result.translation, storage, varName, cache);
-				}
-				_this.callNextAction(cache);
-			});
-	},
-
-	//---------------------------------------------------------------------
-	// Action Bot Mod
-	//
-	// Upon initialization of the bot, this code is run. Using the bot's
-	// DBM namespace, one can add/modify existing functions if necessary.
-	// In order to reduce conflictions between mods, be sure to alias
-	// functions you wish to overwrite.
-	//---------------------------------------------------------------------
-
-	mod: function (DBM) {}
-
-}; // End of module
