+++ actions/store_game_info_MOD.js
